+++ src/main/java/com/course/springfood/infrastructure/service/email/FakeEnvioEmailService.java
@@ -1,13 +1,18 @@
 package com.course.springfood.infrastructure.service.email;
 
+import com.course.springfood.domain.service.EnvioEmailService;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 
 @Slf4j
-public class FakeEnvioEmailService extends SmtpEnvioEmailService {
+public class FakeEnvioEmailService implements EnvioEmailService {
+
+    @Autowired
+    private ProcessadorEmailTemplate processadorEmailTemplate;
 
     @Override
     public void enviar(Mensagem mensagem) {
-        String corpo = processarTemplate(mensagem);
+        String corpo = processadorEmailTemplate.processarTemplate(mensagem);
 
         log.info("[FAKE E-MAIL] Para: {}\n{}", mensagem.getDestinatarios(), corpo);
     }
