+++ src/main/java/com/course/springfood/infrastructure/service/email/ProcessadorEmailTemplate.java
@@ -1,0 +1,27 @@
+package com.course.springfood.infrastructure.service.email;
+
+import com.course.springfood.domain.service.EnvioEmailService.Mensagem;
+import freemarker.template.Configuration;
+import freemarker.template.Template;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+import org.springframework.ui.freemarker.FreeMarkerTemplateUtils;
+
+@Component
+public class ProcessadorEmailTemplate {
+
+    @Autowired
+    private Configuration freemarkerConfig;
+
+    protected String processarTemplate(Mensagem mensagem) {
+        try {
+            Template template = freemarkerConfig.getTemplate(mensagem.getCorpo());
+
+            return FreeMarkerTemplateUtils.processTemplateIntoString(
+                    template, mensagem.getVariaveis());
+        } catch (Exception e) {
+            throw new EmailException("Não foi possível montar o template do e-mail", e);
+        }
+    }
+
+}
