+++ src/main/java/com/course/springfood/infrastructure/service/email/SmtpEnvioEmailService.java
@@ -2,12 +2,9 @@ package com.course.springfood.infrastructure.service.email;
 
 import com.course.springfood.core.email.EmailProperties;
 import com.course.springfood.domain.service.EnvioEmailService;
-import freemarker.template.Configuration;
-import freemarker.template.Template;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.MimeMessageHelper;
-import org.springframework.ui.freemarker.FreeMarkerTemplateUtils;
 
 import javax.mail.MessagingException;
 import javax.mail.internet.MimeMessage;
@@ -21,7 +18,7 @@ public class SmtpEnvioEmailService implements EnvioEmailService {
     private EmailProperties emailProperties;
 
     @Autowired
-    private Configuration freemarkerConfig;
+    private ProcessadorEmailTemplate processadorEmailTemplate;
 
     @Override
     public void enviar(Mensagem mensagem) {
@@ -35,7 +32,7 @@ public class SmtpEnvioEmailService implements EnvioEmailService {
     }
 
     protected MimeMessage criarMimeMessage(Mensagem mensagem) throws MessagingException {
-        String corpo = processarTemplate(mensagem);
+        String corpo = processadorEmailTemplate.processarTemplate(mensagem);
 
         MimeMessage mimeMessage = mailSender.createMimeMessage();
 
@@ -46,17 +43,6 @@ public class SmtpEnvioEmailService implements EnvioEmailService {
         helper.setText(corpo, true);
 
         return mimeMessage;
-    }
-
-    protected String processarTemplate(Mensagem mensagem) {
-        try {
-            Template template = freemarkerConfig.getTemplate(mensagem.getCorpo());
-
-            return FreeMarkerTemplateUtils.processTemplateIntoString(
-                    template, mensagem.getVariaveis());
-        } catch (Exception e) {
-            throw new EmailException("Não foi possível montar o template do e-mail", e);
-        }
     }
 
 }
