+++ src/main/resources/application-development.properties
@@ -1,0 +1,18 @@
+# Database
+spring.datasource.username=root
+spring.datasource.password=
+spring.jpa.show-sql=true
+spring.flyway.locations=classpath:db/migration,classpath:db/testdata
+
+# Security
+springfood.jwt.keystore.jks-location=base64:MIIKvQIBAzCCCnYGCSqGSIb3DQEHAaCCCmcEggpjMIIKXzCCBbYGCSqGSIb3DQEHAaCCBacEggWjMIIFnzCCBZsGCyqGSIb3DQEMCgECoIIFQDCCBTwwZgYJKoZIhvcNAQUNMFkwOAYJKoZIhvcNAQUMMCsEFC3y/mOj80jb4LtZpwEteIY1StnpAgInEAIBIDAMBggqhkiG9w0CCQUAMB0GCWCGSAFlAwQBKgQQ5cC7ZOKvs/2xuOAppI/k5gSCBNA1/anPlghghm8PeFAzjtZF95RhqV4N90DfOYnTnXVDlG54eAxA8fCybf9j6kjn1CwlI/9cgd5CRETp5qV208mK2dCQbZpIpwGD0MgPCWY90EbRKSPzKF10cX05xe1EVc5+TAkCzdq3Kp++c0QAp66+rQTgPfv3RxTvWWqj82iY/LUjMPvqnEz7QPcIEWuEf7ULcQQ8hUWhea38/Sfift96GFnWLoSgI8l0nVgXgHVRsao9kNhZtXnfqNuqvgyxbVGkK9sn0m+uKGj+bF+md9SqoV2CCZB6wPdabSV+vfFJwKWtNSi7XwDk7n1HrmjrnjXoZUsSmBA4XAF1eJ4vYjoO7TIZZTbuBD2x+R2cLb/nQ/sppoxyRSm4xqWOc+BEKpU4ZB0GQBPj9jK5Qgr+gZjVia0OD2+LZFvEHQAv1BdtJ+wRzd7onP3BI3w+IcCoowKv4E3RySndcke0AWP9QsbMzCMdZm1UZrkBTBo45m6n1A8bMrkcCOpx0omO9IrOmd5agMb5kLgJ7SzwDKLT8FGiWNEF2ARIU2EKEJRnnuChzrt5728it/g0uS5mMwB1p64hPUGpYc3Pv9cVYIPGv8BFp1TH+EhIcKD2xLnQ5iXLvgIEqBbwUHnbw95w20ibQyxot+4F/BBI7jg9nQYEUA7hSyi2PlxYgVa3PJXU+lKLA3VJyRb8xDSaWNK+CTmw4vnMMzRuamHHMDIMPM/9woCGGYGAzn9LKvA+CBTaD4uY0W3ZJa8Ab6lgCDUQe3ljnUynKs7eJl7X2tzAIt8nOtiGbhJo4KQ9AyKX57IsjhvCCcQdB7XMhCZS7Iva9X2xi8rZKl+19DoKxzA/3ZNeIEfDy+lolo4quDxLYd9/7f8qot7TDruoO+Z0yBA9UDYfjpMbCQJM9vbSHEkA7GnirzDA5hFRJ5P4QBAq7tOxujFM1TKdMBwgpyUa8Q10D1fO4g0QFrwOyf2r4l2+rfZbHWdmOQW3rgr2KbIMlGFs0+2Obas/9Ds9Pz6LM3P/7ohfY2UxcK7rSqegc/4TjuP+vlAw0n+V+U/+DFkYcVdh/q05k3Ua/AVVzA3r9TLjlnmaC5UC0KtKTGJJTR4yCK9wnaCDVnN8nbOZa+yTpejRT9jRHMhoZdUXa9CwToBdGk9QKGKJvWt966OcLGQmwhM2T7cRYC7F8aFNqS4FPZIHwKe+zcX/5Rxd2Wy/es16eynxuApz2SSK5iLIk+oNOAxzL0TkH/bRMi5n6t2exvcsPS8aQmKjygZ7FZKM9lIGWKGeSfkq3aGsvyyfXa1dxuR7vE/Le8//4vptkeHc5+7VShAYwQVYmSxqJSoExehC9URy7neYipTd9vTJ/59sbEDhnKzWnJC4OPRuavX2730MSsn8TBWBI3JaW+b6E49e4d2jCy4uFVHBZXaoSidXijpavnT7ccoT7ScvJ8dayc/m8JkkojZhIADJQST5RGZoorDOmWxDrusNrGniJR05oAJc/J2fZPsdeDHXzSuICHkgdFhDCChHBeLgUsEgd3KfFiCms6Ba4fjNwNmrXhSq40E1kMFLVZeRkgy2K6yNEPuHOLYkJdHBYExaTlIJEIF2e0co6gJtj1dHP8unB/tM4sqaabt/iAtTGrqOMMYT1qyq7Y15mzFIMCMGCSqGSIb3DQEJFDEWHhQAcwBwAHIAaQBuAGcAZgBvAG8AZDAhBgkqhkiG9w0BCRUxFAQSVGltZSAxNjQxMTcyOTU5NzIzMIIEoQYJKoZIhvcNAQcGoIIEkjCCBI4CAQAwggSHBgkqhkiG9w0BBwEwZgYJKoZIhvcNAQUNMFkwOAYJKoZIhvcNAQUMMCsEFOtNRELOeH7jdqLq3mnEYe9AiR2eAgInEAIBIDAMBggqhkiG9w0CCQUAMB0GCWCGSAFlAwQBKgQQKGTe218JAjYehzHlm7+r4YCCBBBbiTEv2M+EeqjOzMWUo0l1wpTD/yakZDpRBsbCwKx1iVQzGjc5l0W1WJtWbnOpQX70qqCE7SRL4Q5YKGG1TXO8I4+W7SPpoXkD1usZkYpQebqCX7qcIC+WVLNciWEnpXrDZ2uAe/HBgx9QhgVXEVLi1vGw6+0CcEvH1gy5MpWnMwtvpnVeF16Bm2GGFAcnHMe525CFCX0ngm0ziQmhr+RCzKFOLdIEn4qVX4pgBkR6eqX6i3J5sqlTmW8zANshesijAJ6/bz4BvaURk0ePHMDM0sozcLJTrIjwjFzg3U2EIhhFaiuDCCCKcBxgjZw9JQskRuTEb8FnKwDvh+3bgmjRQW0RJ335talX5ZNXZ2Ka5hT1fJTbr4HJ0RChIy21Np/c7ORDa8nrJ6/Or6keGgQL4DDalXtR1YSEp7zcFDpmJW9MGn+qttdisavGaNmETEWkAFb3WMA4wsyvJtcjd36c3X6kzx03+J5ePr6CTuQZJopthXDmoVAdOBA9cDrndFYFn8NNrzgluFrew/BSp0WWp00oC8KRhKSWeiwnQY6/qQFIqPh/4S52s+4fqPvV7eS5QpkMFryjeAmkvce/a5kitEtP6UjcZUjgkroHoCRAzh0uHVA6iO9kBGznned2MujPsQ9T9I+Ztpojf7HOp92GkIDdEPpsN57z+e7Yb/jOi8pYZU1EC4Y4qrh03RZLqp6mE8vv4XKUJRkAi8xwilXA4ndJ1uv5ARKS3ZlciyaV8lZPmDr1wS7YXdQNb3+jSSgZ4uEqnk7Ac9eUXDNq4ue5n8cd5+F0w8iYx36nyb5wd73BYcHkxYPIqH52D5Vdem/sS8WkSF/dRY1iqTEoX7osPaS5YURr61qXNTkBqk/dAo18gGs90ABOvn8s/JcHwzDRMD5DNXlynoYykoCti6Orpf7KzgH9aF+9mFAXdUY2lPF5LUtaFFyFibqFUY2kqyhpTbdOMjdsYu1P0jkVt3Q/YaUZ4aTX2R8bV39FIUPinUbuHvokQYZxW/y0dzs4lzELG3w8V20UliI49LjqMuQEtwv2ykYI0hdw4ykz6Ss2JE9ng6TZb2j/t+FIX1dzhvUUoF7o4dp3oeHoUIIdK+qrHFMaQO8jdqlkYHx6P0fHSQmtKEkSkRhYBMx71/x92vYLENeICDamqTAPYWlmGaJApOQgaL6qXcti/+/yZa1J+0kFn7Cw0t5i6euw+acBLZiAUbvV5q6EyF0c3U15LPRbv0mQpZiFM2o4w/aLKKJwt0AN6Iw0yplMdGfRyN9s4mzOd3ieB2BlISVDiAiCjjClo2dtanG3JBrO2A10wnnzR3709ZODAN3F7aFi51XefMPaRVjFRcHWY8KUd4XhJpekHipm/lkIcaI5ejSE3o1KRDA+MCEwCQYFKw4DAhoFAAQUv+VIDmBY9hWMRz3fsZB8eRYtd6wEFIySGzX81MOkejVBgOtZN1/88TtIAgMBhqA=
+springfood.jwt.keystore.password=123456
+springfood.jwt.keystore.keypair-alias=springfood
+
+# Storage
+springfood.storage.tipo=local
+springfood.storage.local.diretorio-fotos=/home/djalma/Imagens
+
+# Email
+springfood.email.impl=fake
+springfood.email.remetente=SpringFood <enviar@springfood.com>
