+++ src/main/resources/application-production.properties
@@ -1,0 +1,32 @@
+#spring.datasource.username=
+#spring.datasource.password=
+
+# Spring Session (Redis)
+spring.session.store-type=redis
+#spring.redis.host=
+#spring.redis.port=
+#spring.redis.password=
+
+# Logging
+#logging.loggly.token=
+
+# Security
+#springfood.jwt.keystore.jks-location=
+#springfood.jwt.keystore.password=
+#springfood.jwt.keystore.keypair-alias=
+
+# Storage
+springfood.storage.tipo=s3
+springfood.storage.s3.diretorio-fotos=catalogo
+#springfood.storage.s3.id-chave-acesso=
+#springfood.storage.s3.chave-acesso-secreta=
+#springfood.storage.s3.bucket=
+#springfood.storage.s3.regiao=
+
+# Email
+springfood.email.impl=smtp
+#springfood.email.remetente=
+#spring.mail.host=
+#spring.mail.port=
+#spring.mail.username=
+#spring.mail.password=
